move SQL tests to separate `PlanMergeSuite`
